requirements.txt
Django==5.1.3
python-decouple==3.8
djangorestframework==3.15.2
django-cors-headers==4.4.0
pillow==12.1.0
gunicorn==20.1.0
